requirements.txt
fastapi
jinja2
mysql-connector-python
black
uvicorn
python-multipart